refactored CameraX code
missing: Gallery function, flipped image fix
