+++ app/src/main/java/com/example/booktrackerapp/model/service/ImageUri.kt
@@ -1,9 +1,10 @@
+package com.example.booktrackerapp.model.service
+
 import android.net.Uri
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 
-class ImageViewModel : ViewModel() {
+class ImageUri {
     private val _imageUri = MutableLiveData<Uri?>()
     val imageUri: LiveData<Uri?> get() = _imageUri
