+++ app/src/main/java/com/example/booktrackerapp/navigation/Navigation.kt
@@ -1,26 +1,28 @@
 package com.example.booktrackerapp.navigation
 
-import CameraScreen
-import ImageViewModel
 import PreviewScreen
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.booktrackerapp.model.service.ImageUri
+import com.example.booktrackerapp.screens.CameraScreen
 import com.example.booktrackerapp.screens.HomeScreen
 import com.example.booktrackerapp.screens.LibraryScreen
 import com.example.booktrackerapp.screens.SignInScreen
 import com.example.booktrackerapp.screens.SignUpScreen
 import com.example.booktrackerapp.screens.SplashScreen
 import com.example.booktrackerapp.screens.Userscreen
+import com.example.booktrackerapp.viewModel.CameraViewModel
 import com.example.booktrackerapp.viewModel.HomeViewModel
 
 @Composable
 fun Navigation() {
     val navController = rememberNavController()
     val viewModel: HomeViewModel = viewModel()
-    val imageViewModel: ImageViewModel = viewModel()
+    val cameraViewModel: CameraViewModel = viewModel()
+    val imageUriHolder = ImageUri()
 
     NavHost(
         navController = navController,
@@ -47,11 +49,11 @@ fun Navigation() {
         }
         composable(Screen.CameraScreen.route)
         {
-            CameraScreen(navController = navController, imageViewModel = imageViewModel)
+            CameraScreen(navController = navController, cameraViewModel = cameraViewModel, imageUriHolder = imageUriHolder)
         }
         composable(Screen.PreviewScreen.route)
         {
-            PreviewScreen(navController = navController, imageViewModel = imageViewModel)
+            PreviewScreen(navController = navController, imageUriHolder = imageUriHolder)
         }
     }
 }
